% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.WilcoxonMannWhitney.R
\name{legacy.WilcoxonMannWhitney}
\alias{legacy.BrownHettmansperger}
\alias{legacy.JonckheereTerpstra}
\alias{legacy.KruskalWallis}
\alias{legacy.MackSkillings}
\alias{legacy.MackWolfe}
\alias{legacy.WilcoxonMannWhitney}
\title{Non-PIM implementations of distributionfree tests}
\usage{
legacy.WilcoxonMannWhitney(data, out, group, block, verbosity = 0)

legacy.KruskalWallis(data, out, group, block, verbosity = 0)

legacy.MackSkillings(data, out, group, block, verbosity = 0)

legacy.BrownHettmansperger(data, out, group, block, verbosity = 0)

legacy.JonckheereTerpstra(data, out, group, block, verbosity = 0)

legacy.MackWolfe(data, out, group, block, verbosity = 0, levelP)
}
\arguments{
\item{data}{Context where \code{out}, \code{group} and \code{block} are to be interpreted}

\item{out}{Column of \code{data} that holds the outcomes (responses)}

\item{group}{Column of \code{data} that holds the predicting variable}

\item{block}{Column of \code{data} that holds the blocking variable. This can be left out 
if no blocking is present}

\item{verbosity}{The higher this value, the more levels of progress and debug 
information is displayed (note: in R for Windows, turn off buffered output)}

\item{levelP}{The level of the grouping variable that is the "top" of the umbrella.}
}
\value{
a list holding the following items:
\item{statistic}{The test statistic.} 
\item{p.value}{p-value for the test.} 
\item{df}{Degress of freedom (if relevant, otherwise \code{NA}).} 
\item{conversion}{Function to convert the pim estimates to the test statistic.}
}
\description{
Non-PIM implementations of distributionfree tests
}
\note{
These functions are merely provided for comparison. The pim function supports
	generic ways of estimating these values. For generosity, the parameters to all of
	them are the same, though some are not relevant (e.g. \code{block} for WMW).
}
\keyword{fit}
\keyword{legacy}
\keyword{pim}

