% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pim-package.R
\docType{package}
\name{pimold-package}
\alias{pimold-package}
\title{Probabilistic Index Models}
\description{
Fit a probabilistic index model. This package is the original
one described in the dissertation of prof. dr. Jan De Neve and the paper
by De Neve and Thas (2015): "A regression framework for rank tests based on the 
probabilistic index model", JASA 511, 1276-1283. Note that the complete pim
package is reworked to have a more intuitive formula interface. This old
version exists only for reference to those papers.
}
\details{
\tabular{ll}{ Package: \tab pimold\cr Type: \tab Package\cr Version: \tab
1.1.5.6\cr Date: \tab 2014-10-16\cr License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr Depends: R (>= 2.12), Matrix, nleqslv, MASS\cr Suggests: glmnet (>= 1.9-3), BB, lqa}
}
\author{
Nick Sabbe \email{Nick.Sabbe@UGent.be} Jan De Neve \email{JanR.DeNeve@UGent.be} Joris Meys \email{Joris.Meys@UGent.be}
}
\references{
[PENDING]
}
\keyword{package}

