% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.pim.R
\name{print.pim}
\alias{print.pim}
\title{Display a PIM}
\usage{
\method{print}{pim}(x,...)
}
\arguments{
\item{x}{\code{\link{pim}} object.}

\item{\dots}{Ignored currently.}
}
\value{
Nothing.
}
\description{
Display a PIM
}

